refactored standard lib down to a file, also updated __init__
