+++ DiceStatistics/__init__.py
@@ -4,6 +4,5 @@ __all__ = []
 from DiceStatistics.Dice import Dice 
 from DiceStatistics.Pool import Pool
 from DiceStatistics.interface import d, z
-from DiceStatistics._parser import faces_to_prop, text_to_faces
-from DiceStatistics.systems.standard import d6,d8,d10,d12,d20,d100,z9
+from DiceStatistics.standard import d6,d8,d10,d12,d20,d100,z9
 from DiceStatistics._math import ceil, floor
