+++ DiceStatistics/standard.py
@@ -9,3 +9,5 @@ d20  = d(20)
 d100 = d(100)
 
 z9 = z(9)
+
+__all__ = [d4, d6, d8, d10, d12, d20, d100, z9]
