--- DiceStatistics/systems/__init__.py
@@ -1,1 +1,0 @@
-from DiceStatistics.systems.standard import d4,d6,d8,d10,d12,d20,d100,z9
